Moving profiling to it's own trait
